requirements.txt
boto3==1.24.67
botocore==1.27.67
jmespath==1.0.1
passlib==1.7.4
PyJWT==2.4.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
fpdf2==2.6.1
pandas
numpy
matplotlib
borb
PyPDF4
reportlab
pdfminer